requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
amqp>=5.0.6
    # via kombu
asgiref>=3.4.1
    # via django
backcall>=0.2.0
    # via ipython
billiard>=3.6.4.0
    # via celery
brotlipy>=0.7.0
    # via -r requirements.in
cached-property>=1.5.2
    # via kombu
celery[redis]>=5.1.2
    # via -r requirements.in
certifi>=2021.5.30
    # via
    #   requests
    #   sentry-sdk
cffi>=1.14.6
    # via
    #   brotlipy
    #   cryptography
charset-normalizer>=2.0.2
    # via requests
click>=7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean>=0.0.3
    # via celery
click-plugins>=1.1.1
    # via celery
click-repl>=0.2.0
    # via celery
cryptography>=3.4.7
    # via social-auth-core
decorator>=5.0.9
    # via
    #   ipython
    #   traitlets
defusedxml>=0.7.1
    # via
    #   odfpy
    #   python3-openid
    #   social-auth-core
diff-match-patch>=20200713
    # via django-import-export
dj-database-url>=0.5.0
    # via -r requirements.in
django>=3.2.5
    # via
    #   -r requirements.in
    #   django-import-export
    #   django-js-reverse
    #   django-log-request-id
    #   django-model-utils
    #   django-render-block
    #   djangorestframework
django-import-export>=2.5.0
    # via -r requirements.in
django-js-reverse>=0.9.1
    # via -r requirements.in
django-log-request-id>=1.6.0
    # via -r requirements.in
django-model-utils>=4.1.1
    # via -r requirements.in
django-render-block>=0.8.1
    # via django-templated-email
django-templated-email>=3.0.0
    # via -r requirements.in
django-webpack-loader>=1.1.0
    # via -r requirements.in
djangorestframework>=3.12.4
    # via -r requirements.in
et-xmlfile>=1.1.0
    # via openpyxl
gunicorn>=20.1.0
    # via -r requirements.in
idna>=3.2
    # via requests
importlib-metadata>=4.6.1
    # via kombu
ipython>=7.16.1
    # via -r requirements.in
ipython-genutils>=0.2.0
    # via traitlets
jedi>=0.18.0
    # via ipython
kombu>=5.1.0
    # via celery
markuppy>=1.14
    # via tablib
oauthlib>=3.1.1
    # via
    #   requests-oauthlib
    #   social-auth-core
odfpy>=1.4.1
    # via tablib
openpyxl>=3.0.7
    # via tablib
parso>=0.8.2
    # via jedi
pexpect>=4.8.0
    # via ipython
pickleshare>=0.7.5
    # via ipython
prompt-toolkit>=3.0.19
    # via
    #   click-repl
    #   ipython
psutil>=5.8.0
    # via -r requirements.in
psycopg2>=2.9.1
    # via -r requirements.in
ptyprocess>=0.7.0
    # via pexpect
pycparser>=2.20
    # via cffi
pygments>=2.9.0
    # via ipython
pyjwt>=2.1.0
    # via social-auth-core
python-decouple>=3.4
    # via -r requirements.in
python3-openid>=3.2.0
    # via social-auth-core
pytz>=2021.1
    # via
    #   celery
    #   django
pyyaml>=5.4.1
    # via tablib
redis>=3.5.3
    # via celery
requests>=2.26.0
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib>=1.3.0
    # via social-auth-core
sentry-sdk>=1.3.0
    # via -r requirements.in
six>=1.16.0
    # via
    #   click-repl
    #   social-auth-app-django
    #   traitlets
social-auth-app-django>=4.0.0
    # via -r requirements.in
social-auth-core>=4.1.0
    # via social-auth-app-django
sqlparse>=0.4.1
    # via django
tablib[html,ods,xls,xlsx,yaml]>=3.0.0
    # via django-import-export
traitlets>=4.3.3
    # via ipython
typing-extensions>=3.10.0.0
    # via
    #   asgiref
    #   importlib-metadata
urllib3>=1.26.6
    # via
    #   requests
    #   sentry-sdk
vine>=5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth>=0.2.5
    # via prompt-toolkit
whitenoise>=5.2.0
    # via -r requirements.in
xlrd>=2.0.1
    # via tablib
xlwt>=1.3.0
    # via tablib
zipp>=3.5.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
